ai/download_sp/free-video-downloader/backend/requirements.txt
fastapi>=0.135.0
uvicorn>=0.34.0
yt-dlp>=2026.1.0
python-multipart>=0.0.18
httpx>=0.28.0
openai>=1.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
stripe>=8.0.0
python-dateutil>=2.8.0
